refactor(organigram):helper storage
