+++ frontend/src/app/routes/organigrama/organigramaRamas_Subramas/services/image-upload-core.service.ts
@@ -1,5 +1,5 @@
 import api from '@/api/axios';
-import { postFormData } from '@/api/formData';
+import { uploadToStorage } from '@/api/upload';
 import { sectionPath } from '@/api/organigramaApi';
 
 type MaybeAxiosError = { response?: { data?: unknown } };
@@ -29,7 +29,7 @@ export const diagnoseBatchImageUpload = async (
     const formData = new FormData();
     formData.append('file', file);
     
-  const uploadResponse = await postFormData<UploadResponse>('/storage/upload', formData);
+  const uploadResponse = await uploadToStorage<UploadResponse>(formData);
     console.log('✅ [DIAGNÓSTICO] Upload exitoso, objectId:', uploadResponse.objectId);
 
     // Paso 2: Agregar a galería usando PATCH
@@ -92,7 +92,7 @@ export const uploadSectionIcon = async (
     formData.append('file', file);
     
     console.log('🔄 [ImageUploadService] Subiendo archivo al storage...');
-  const uploadResponse = await postFormData<UploadResponse>('/storage/upload', formData);
+  const uploadResponse = await uploadToStorage<UploadResponse>(formData);
     console.log('✅ [ImageUploadService] Archivo subido, objectId:', uploadResponse.objectId);
     
     // Paso 2: Usar endpoint PATCH específico para icono
@@ -171,7 +171,7 @@ export const uploadSectionMainImage = async (
     const formData = new FormData();
     formData.append('file', file);
     
-  const uploadResponse = await postFormData<UploadResponse>('/storage/upload', formData);
+  const uploadResponse = await uploadToStorage<UploadResponse>(formData);
     
     // Paso 2: Usar endpoint PATCH específico para imagen principal
     console.log('🔄 [ImageUploadService] Asociando imagen principal usando endpoint PATCH específico...');
@@ -248,7 +248,7 @@ export const uploadGalleryImages = async (
       const formData = new FormData();
       formData.append('file', file);
       
-  const uploadResponse = await postFormData<UploadResponse>('/storage/upload', formData);
+  const uploadResponse = await uploadToStorage<UploadResponse>(formData);
       objectIds.push(uploadResponse.objectId);
       urls.push(uploadResponse.url || uploadResponse.objectId);
     }
